backend/pyproject.toml
[project]
name = "wearefriends"
version = "0.1.0"
description = ""
authors = [
    {name = "yudai-kdix",email = "yudai.homma222@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "openai (>=1.78.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "gtts (>=2.5.4,<3.0.0)",
    "ultralytics (>=8.3.134,<9.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false